inline global_reboot_test
