+++ Cargo.toml
@@ -363,6 +363,7 @@ members = [
     "rs/tests/consensus/utils",
     "rs/tests/driver",
     "rs/tests/execution",
+    "rs/tests/message_routing",
     "rs/tests/httpbin-rs",
     "rs/tests/networking",
     "rs/tests/networking/canisters",
