+++ rs/tests/src/message_routing/mod.rs
@@ -1,5 +1,4 @@
 pub mod compatibility;
-pub mod global_reboot_test;
 pub mod malicious_slices;
 pub mod memory_safety_test;
 pub mod rejoin_test;
